fix(ci): line-limit — extract workspaces_retry recovery helpers over 1500 lines
